better CtrlZ and ready Search
