+++ src/main/java/com/mit/blocks/renderable/FactoryRenderableBlock.java
@@ -138,4 +138,5 @@ public class FactoryRenderableBlock extends RenderableBlock implements Cloneable
     {
         super.setZoomLevel(newZoom);
     }
+    
 }
